Fixed Weapon Switch  Started WeaponTracker V1
Weapon Switch now no longer needs a specified size

Weapon Tracker  tracks current ammo for all guns.  Script is set up to only hold 1 of each weapon type.  Weapons can be switched to only when that weapon has ammo.  

Idea: if a weapon has ammo it is display in INV.  if it does not have ammo it will be discarded until ammo is found.